refactor: SuggestionsRepo
