+++ engine/language-server/src/test/scala/org/enso/languageserver/runtime/SuggestionsDatabaseEventsListenerTest.scala
@@ -17,15 +17,14 @@ import org.enso.languageserver.data.{
 import org.enso.languageserver.session.JsonSession
 import org.enso.languageserver.session.SessionRouter.DeliverToJsonController
 import org.enso.polyglot.runtime.Runtime.Api
-import org.enso.searcher.sql.{SqlDatabase, SqlSuggestionsRepo}
-import org.enso.searcher.{Database, SuggestionsRepo}
+import org.enso.searcher.SuggestionsRepo
+import org.enso.searcher.sql.SqlSuggestionsRepo
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
-import slick.dbio.DBIO
 
-import scala.concurrent.{Await, Future}
 import scala.concurrent.duration._
+import scala.concurrent.{Await, Future}
 
 class SuggestionsDatabaseEventsListenerTest
     extends TestKit(ActorSystem("TestSystem"))
@@ -46,8 +45,8 @@ class SuggestionsDatabaseEventsListenerTest
   "SuggestionsHandler" should {
 
     "subscribe to notification updates" taggedAs Retry() in withDb {
-      (router, repo, db) =>
-        val handler  = newEventsListener(router.ref, repo, db)
+      (router, repo) =>
+        val handler  = newEventsListener(router.ref, repo)
         val clientId = UUID.randomUUID()
 
         handler ! AcquireCapability(
@@ -57,8 +56,8 @@ class SuggestionsDatabaseEventsListenerTest
         expectMsg(CapabilityAcquired)
     }
 
-    "receive runtime updates" taggedAs Retry() in withDb { (router, repo, db) =>
-      val handler  = newEventsListener(router.ref, repo, db)
+    "receive runtime updates" taggedAs Retry() in withDb { (router, repo) =>
+      val handler  = newEventsListener(router.ref, repo)
       val clientId = UUID.randomUUID()
 
       // acquire capability
@@ -85,13 +84,13 @@ class SuggestionsDatabaseEventsListenerTest
       )
 
       // check database entries exist
-      val records = Await.result(db.run(repo.getAll), Timeout)
+      val (_, records) = Await.result(repo.getAll, Timeout)
       records.map(_.suggestion) should contain theSameElementsAs Suggestions.all
     }
 
     "apply runtime updates in correct order" taggedAs Retry() in withDb {
-      (router, repo, db) =>
-        val handler  = newEventsListener(router.ref, repo, db)
+      (router, repo) =>
+        val handler  = newEventsListener(router.ref, repo)
         val clientId = UUID.randomUUID()
 
         // acquire capability
@@ -119,35 +118,31 @@ class SuggestionsDatabaseEventsListenerTest
         )
 
         // check that database entries removed
-        val records = Await.result(db.run(repo.getAll), Timeout)
-        records.map(
-          _.suggestion
-        ) should contain theSameElementsAs Suggestions.all
+        val (_, all) = Await.result(repo.getAll, Timeout)
+        all.map(_.suggestion) should contain theSameElementsAs Suggestions.all
     }
 
   }
 
   def newEventsListener(
     router: ActorRef,
-    repo: SuggestionsRepo[DBIO],
-    db: Database[DBIO, Future]
+    repo: SuggestionsRepo[Future]
   ): ActorRef =
-    system.actorOf(SuggestionsDatabaseEventsListener.props(router, repo, db))
+    system.actorOf(SuggestionsDatabaseEventsListener.props(router, repo))
 
   def newJsonSession(clientId: UUID): JsonSession =
     JsonSession(clientId, TestProbe().ref)
 
   def withDb(
-    test: (TestProbe, SuggestionsRepo[DBIO], Database[DBIO, Future]) => Any
+    test: (TestProbe, SuggestionsRepo[Future]) => Any
   ): Unit = {
     val dbPath = Files.createTempFile("suggestions", ".db")
     system.registerOnTermination(Files.deleteIfExists(dbPath))
-    val repo = new SqlSuggestionsRepo()
-    val db   = new SqlDatabase()
-    Await.ready(db.run(repo.init), Timeout)
+    val repo = SqlSuggestionsRepo()
+    Await.ready(repo.init, Timeout)
     val router = TestProbe("sessionRouterProbe")
 
-    try test(router, repo, db)
-    finally db.close()
+    try test(router, repo)
+    finally repo.close()
   }
 }
